Prepare for Compact - Simplify Default vs Constant Serializers
