+++ src/Hazelcast.Net.Tests/Aggregating/AggregatingTests.cs
@@ -17,7 +17,6 @@ using Hazelcast.Aggregation;
 using Hazelcast.Core;
 using Hazelcast.Serialization;
 using Hazelcast.Serialization.ConstantSerializers;
-using Hazelcast.Serialization.DefaultSerializers;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
 
@@ -33,7 +32,6 @@ namespace Hazelcast.Tests.Aggregating
         {
             _serializationService = new SerializationServiceBuilder(new NullLoggerFactory())
                 .AddDefinitions(new ConstantSerializerDefinitions()) // use constant serializers not CLR serialization
-                .AddDefinitions(new DefaultSerializerDefinitions()) // same
                 .AddHook<AggregatorDataSerializerHook>()
                 .Build();
         }
