+++ src/Hazelcast.Net.Tests/Protocol/HoldersTests.cs
@@ -19,7 +19,6 @@ using Hazelcast.Protocol.Models;
 using Hazelcast.Query;
 using Hazelcast.Serialization;
 using Hazelcast.Serialization.ConstantSerializers;
-using Hazelcast.Serialization.DefaultSerializers;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
 
@@ -33,7 +32,6 @@ namespace Hazelcast.Tests.Protocol
         {
             var serializationService = new SerializationServiceBuilder(new NullLoggerFactory())
                 .AddDefinitions(new ConstantSerializerDefinitions()) // use constant serializers not CLR serialization
-                .AddDefinitions(new DefaultSerializerDefinitions()) // same
                 .Build();
 
             var holder = GetAnchorDataListHolder(serializationService, out var pageList, out var dataList);
@@ -71,7 +69,6 @@ namespace Hazelcast.Tests.Protocol
         {
             var serializationService = new SerializationServiceBuilder(new NullLoggerFactory())
                 .AddDefinitions(new ConstantSerializerDefinitions()) // use constant serializers not CLR serialization
-                .AddDefinitions(new DefaultSerializerDefinitions()) // same
                 .AddHook<PredicateDataSerializerHook>()
                 .Build();
