+++ src/Hazelcast.Net.Tests/Query/PredicatesTests.cs
@@ -20,7 +20,6 @@ using Hazelcast.Core;
 using Hazelcast.Query;
 using Hazelcast.Serialization;
 using Hazelcast.Serialization.ConstantSerializers;
-using Hazelcast.Serialization.DefaultSerializers;
 using Hazelcast.Testing.Predicates;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
@@ -37,7 +36,6 @@ namespace Hazelcast.Tests.Query
         {
             _serializationService = new SerializationServiceBuilder(new NullLoggerFactory())
                 .AddDefinitions(new ConstantSerializerDefinitions()) // use constant serializers not CLR serialization
-                .AddDefinitions(new DefaultSerializerDefinitions()) // same
                 .AddHook<PredicateDataSerializerHook>()
                 .Build();
         }
