+++ src/Hazelcast.Net.Tests/Serialization/ConstantSerializerTest.cs
@@ -18,7 +18,6 @@ using System.Numerics;
 using Hazelcast.Core;
 using Hazelcast.Serialization;
 using Hazelcast.Serialization.ConstantSerializers;
-using Hazelcast.Serialization.DefaultSerializers;
 using Hazelcast.Testing;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
@@ -42,7 +41,6 @@ namespace Hazelcast.Tests.Serialization
                 .SetEndianness(endianness)
                 .SetPortableVersion(1)
                 .AddDefinitions(new ConstantSerializerDefinitions())
-                .AddDefinitions(new DefaultSerializerDefinitions())
                 .Build();
         }
