+++ src/Hazelcast.Net.Tests/Serialization/JsonValueSerializationTest.cs
@@ -16,7 +16,6 @@ using System;
 using Hazelcast.Core;
 using Hazelcast.Serialization;
 using Hazelcast.Serialization.ConstantSerializers;
-using Hazelcast.Serialization.DefaultSerializers;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
 
@@ -32,7 +31,6 @@ namespace Hazelcast.Tests.Serialization
         {
             _serializationService = new SerializationServiceBuilder(new NullLoggerFactory())
                 .AddDefinitions(new ConstantSerializerDefinitions())
-                .AddDefinitions(new DefaultSerializerDefinitions())
                 .Build();
         }
