+++ src/Hazelcast.Net/Serialization/ConstantSerializers/ArrayStreamSerializer.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Hazelcast.Serialization.DefaultSerializers
+namespace Hazelcast.Serialization.ConstantSerializers
 {
     internal class ArrayStreamSerializer : SingletonSerializerBase<object[]>
     {
