+++ src/Hazelcast.Net/Serialization/ConstantSerializers/HBigDecimalSerializer.cs
@@ -16,7 +16,7 @@ using System;
 using System.Numerics;
 using Hazelcast.Models;
 
-namespace Hazelcast.Serialization.DefaultSerializers
+namespace Hazelcast.Serialization.ConstantSerializers
 {
     internal class HBigDecimalSerializer : SingletonSerializerBase<HBigDecimal>
     {
