+++ src/Hazelcast.Net/Serialization/ConstantSerializers/ListSerializer.cs
@@ -15,7 +15,7 @@
 using System.Collections.Generic;
 using Hazelcast.Core;
 
-namespace Hazelcast.Serialization.DefaultSerializers
+namespace Hazelcast.Serialization.ConstantSerializers
 {
     internal class ListSerializer<T> : SingletonSerializerBase<List<T>>
     {
