--- src/Hazelcast.Net/Serialization/DefaultSerializers/HLocalDateSerializer.cs
@@ -1,39 +1,0 @@
-﻿// Copyright (c) 2008-2022, Hazelcast, Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Hazelcast.Models;
-
-namespace Hazelcast.Serialization.DefaultSerializers
-{
-    internal class HLocalDateSerializer : SingletonSerializerBase<HLocalDate>
-    {
-        public override int TypeId => SerializationConstants.JavaDefaultTypeLocalDate;
-
-        public override HLocalDate Read(IObjectDataInput input)
-        {
-            var year = input.ReadInt();
-            var month = input.ReadByte();
-            var date = input.ReadByte();
-
-            return new HLocalDate(year, month, date);
-        }
-
-        public override void Write(IObjectDataOutput output, HLocalDate obj)
-        {
-            output.WriteInt(obj.Year);
-            output.WriteByte(obj.Month);
-            output.WriteByte(obj.Day);
-        }
-    }
-}
